detach Data annotation classes and modify associated Controllers
