+++ Project-Mug/src/main/java/recyclemug/ProjectMug/api/CustomerApiController.java
@@ -5,6 +5,8 @@ import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
+import recyclemug.ProjectMug.data.CreateCustomerRequest;
+import recyclemug.ProjectMug.data.CreateCustomerResponse;
 import recyclemug.ProjectMug.domain.user.Customer;
 import recyclemug.ProjectMug.service.CustomerService;
 
@@ -34,22 +36,5 @@ public class CustomerApiController {
         return new CreateCustomerResponse("success", "회원가입에 성공했습니다.");
     }
 
-    @Data
-    static class CreateCustomerRequest {
-        private String role; // "customer" or "partner"
-        private String id; 	// String (email)
-        private String pw;		// String
-        private String tel;		// String
-        private String address;	// String (partner일경우 "")
-        private String address_num;	// String (5자리숫자, partner일경우 "")
-        private String address_detail;
-    }
-
-    @Data
-    @AllArgsConstructor
-    static class CreateCustomerResponse {
-        private String result;
-        private String message;
-    }
 
 }
