+++ Project-Mug/src/main/java/recyclemug/ProjectMug/data/CreateCustomerResponse.java
@@ -1,0 +1,14 @@
+package recyclemug.ProjectMug.data;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+/**
+ * Customer 반환 클래스
+ */
+@Data
+@AllArgsConstructor
+public class CreateCustomerResponse {
+    private String result;
+    private String message;
+}
