+++ Project-Mug/src/main/java/recyclemug/ProjectMug/data/CreatePartnerResponse.java
@@ -1,0 +1,14 @@
+package recyclemug.ProjectMug.data;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+/**
+ * Partner 반환 클래스
+ */
+@Data
+@AllArgsConstructor
+public class CreatePartnerResponse {
+    private String result;
+    private String message;
+}
